portfolios-desarrollo/chatbot-dev/backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
google-generativeai==0.5.4
langchain==0.1.17
langchain-google-genai==1.0.3
chromadb==0.4.15
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0